Create an Order domain model
The Order model lives in the 'features/orders/domain' folder. The folder
structure is meant to mirror a design similar to the Clean Architecture
to ensure a separation of concerns between the UI code, the business
logic and the data access logic.

A 'feature' folder will be split into 4 directories/layers:
    * 'presentation' layer for UI code,
    * 'application' layer for services,
    * 'domain' layer for modelling the data,
    * 'data' layer for data sources

So far, we've only defined the 'domain' layer, the other layers will be
created later on.