Emulate fetching data w/ FakeOrdersRepository
A FakeOrdersRepository is created to emulate fetching the orders data
from an access point. For now, this class will serve for sketching the
UI logic, and can later be used for writing tests.

The riverpod v2.3.7 package was installed, so that the orders data could
be exposed to the widgets via a FutureProvider.